+++ R/utils_gt.R
@@ -55,7 +55,7 @@ hf_to_gt_tbl <- function(gt, header, subheader, footer){
       old_sub <- gt$`_heading`$subtitle
       # add subtitle text on top of groups if exists
       if(length(subheader)>0){
-        subtitle = gt::md(paste0(paste0(subheader, collapse = "<br>"), "<br>",old_sub))
+        subtitle = gt::md(paste0(paste0(subheader, collapse = "<br>"), "<br>",paste0(old_sub, collapse = "<br>")))
       }else{
         subtitle = gt::md(old_sub)
       }
@@ -169,4 +169,145 @@ apply_to_grp <- function(func, args, call = rlang::caller_env()){
   }
 
   gt_obj
+}
+
+#' Convert png object to gt from docorator object
+#' @param x docorator object
+#' @export
+#' @keywords internal
+png_to_gt <- function(x){
+
+  # save png to temp file
+ temp_png <- tempfile(
+    pattern = "temp_png_",
+    tmpdir = tempdir(),
+    fileext = ".png")
+
+ png::writePNG(x$display$png, temp_png)
+
+  # convert to gt
+  gt <- dplyr::tibble(ggplot =  temp_png) |>
+    gt::gt() |>
+    gt::fmt_image(columns = dplyr::everything(), sep = ",", width = "6in") |>
+    # remove column headers and borders
+    gt::tab_options(
+      column_labels.hidden = TRUE,
+      table.border.top.style = "hidden",
+      table.border.bottom.style = "hidden"
+    )
+
+  gt
+
+}
+
+#' Convert ggplot object to gt from docorator object
+#' @param x docorator object
+#' @export
+#' @keywords internal
+gg_to_gt <- function(x){
+
+  if (!inherits(x$display, "ggplot")) {
+    cli::cli_abort("The display must be class `ggplot`, not {.obj_type_friendly {x$display}}.",
+                   call = rlang::caller_env())
+  }
+
+  # remove header footer information
+  display_info <- hf_strip(x$display)
+
+  # convert to gt
+  gt <- dplyr::tibble(ggplot =  display_info$display |>
+                 gg_to_image(fig_dim = x$fig_dim, path = tempdir())) |>
+    gt::gt() |>
+    gt::fmt_image(columns = dplyr::everything(), sep = ",", width = "6in") |>
+    # add header and footer information
+    gt::tab_header(title = display_info$head_data, subtitle = display_info$subhead_data) |>
+    # remove column headers and borders
+    gt::tab_options(
+      column_labels.hidden = TRUE,
+      table.border.top.style = "hidden",
+      table.border.bottom.style = "hidden"
+    )
+
+  # add footnotes if not NULL - avoids null row being added
+  if(!is.null(display_info$foot_data)){
+    gt <- gt |>
+      gt::tab_footnote(footnote = display_info$foot_data)
+  }
+
+  gt
+
+}
+
+#' ggplot to image filepath
+#' Taken from code here: https://github.com/rstudio/gt/blob/6312cf5be92fae41633b5df7d41aa948b850aaf8/R/image.R#L351C1-L392C2
+#' @param plot_object ggplot object
+#' @param fig_dim vector containing figure height and width in inches. Defaults to c(5,8) - docorator defaults
+#' @param path file path including the png location to save the output to
+#' @return filepath to snap of ggplot
+#' @keywords internal
+gg_to_image <- function(plot_object, fig_dim = c(5,8), path = getwd()) {
+
+
+  # Upgrade x to a list if only a single ggplot object is provided
+  if (inherits(plot_object, "gg")) {
+    plot_object <- list(plot_object)
+  }
+
+  vapply(
+    seq_along(plot_object),
+    FUN.VALUE = character(1L),
+    USE.NAMES = FALSE,
+    FUN = function(x) {
+
+      filename <- tempfile(
+        pattern = "temp_gt_ggplot_",
+        tmpdir = path,
+        fileext = ".png")
+
+      # Save PNG file to disk
+      ggplot2::ggsave(
+        filename = filename,
+        create.dir = TRUE,
+        plot = plot_object[[x]],
+        device = "png",
+        dpi = 100,
+        width = fig_dim[2],
+        height = fig_dim[1],
+        units = "in"
+      )
+
+      filename
+
+    }
+  )
+}
+
+#' strip header footer info from ggplot object
+#' @param x a ggplot object
+#'
+#' @return list of header footer info
+#'
+#' @export
+#' @keywords internal
+hf_strip <- function(x){
+
+
+  if (!inherits(x, "ggplot")) {
+    cli::cli_abort("The display must be class `ggplot`, not {.obj_type_friendly {x}}.",
+                   call = rlang::caller_env())
+  }
+
+  head_data <- x$labels$title
+  # move tag to subtitle
+  subhead_data <- c(x$labels$subtitle,x$labels$tag)
+  foot_data <- x$labels$caption
+
+  # set to null in ggplot object
+  x$labels[c("title","subtitle", "caption", "tag")] <- NULL
+
+  list(display = x,
+       head_data = head_data,
+       subhead_data = subhead_data,
+       foot_data = foot_data)
+
 }