+++ tests/testthat/test-render.R
@@ -199,6 +199,66 @@ test_that("render to pdf, lists - quarto", {
   unlink(temp_dir_path, recursive = TRUE)
 })
 
+test_that("render to pdf - transform", {
+
+  skip_on_cran()
+  skip_on_ci()
+
+  my_gt <- gt::exibble |>
+    gt::gt(
+      rowname_col = "row",
+      groupname_col = "group"
+    )
+
+  transform <- function(x){stringr::str_replace_all(x,"num", "NUM")}
+
+  temp_dir_path <- file.path(rprojroot::is_testthat$find_file(), "tempdir")
+  if (!dir.exists(temp_dir_path)){
+    dir.create(file.path(rprojroot::is_testthat$find_file(), "tempdir"))
+  }
+
+  docorator <- as_docorator(
+    x = my_gt,
+    header = fancyhead(fancyrow("first line header"), fancyrow("second line header")),
+    footer = NULL,
+    display_name = "my_first_gt",
+    display_loc = temp_dir_path,
+    save_object = FALSE
+  )
+
+  res <- suppressMessages( docorator |> render_pdf(transform = transform)
+  )
+
+  expect_true(file.exists(file.path(temp_dir_path, "my_first_gt.pdf")))
+  expect_true(stringr::str_detect(pdftools::pdf_text(file.path(temp_dir_path, "my_first_gt.pdf")),"NUM"))
+  expect_false(stringr::str_detect(pdftools::pdf_text(file.path(temp_dir_path, "my_first_gt.pdf")),"num"))
+
+  # quarto render
+  temp_dir_path2 <- file.path(rprojroot::is_testthat$find_file(), "tempdir2")
+  if (!dir.exists(temp_dir_path2)){
+    dir.create(file.path(rprojroot::is_testthat$find_file(), "tempdir2"))
+  }
+
+  withr::with_dir(
+    file.path(rprojroot::is_testthat$find_file(), "tempdir2"),
+    code = {
+      docorator <- as_docorator(
+        x = my_gt,
+        header = fancyhead(fancyrow("first line header"), fancyrow("second line header")),
+        footer = NULL,
+        display_name = "my_first_gt",
+        save_object = FALSE
+      )
+      expect_warning(suppressMessages( docorator |> render_pdf(
+        quarto = TRUE, transform = transform
+      )))
+    })
+  expect_true(file.exists(file.path(temp_dir_path2, "my_first_gt.pdf")))
+
+  unlink(temp_dir_path, recursive = TRUE)
+  unlink(temp_dir_path2, recursive = TRUE)
+})
+
 test_that("render to rtf works", {
 
   skip_on_cran()
@@ -318,6 +378,58 @@ test_that("rtf headers",{
   res2 <- suppressMessages( docorator |> render_rtf(use_page_header = TRUE))
   doc2 <- readLines(file.path(temp_dir_path, "my_gt_header.rtf"))|> paste0(collapse="")
   expect_true(grepl("\\{\\\\header",doc2))
+
+  unlink(temp_dir_path, recursive = TRUE)
+})
+
+test_that("render to rtf, lists of figures", {
+
+  skip_on_cran()
+  skip_on_ci()
+
+  png_obj1 <- png_path(path = system.file("extdata/test_image.png", package = "docorator"))
+  png_obj2 <- png_path(path = system.file("extdata/test_image.png", package = "docorator"))
+
+  ggplot1 <- ggplot2::ggplot(data = mtcars, ggplot2::aes(y=cyl, x=mpg)) +
+    ggplot2::geom_point() +
+    ggplot2::labs(title = "title1", subtitle = "subtitle1", tag = "tag1", caption = "footnote1")
+  ggplot2 <- ggplot2::ggplot(data = mtcars, ggplot2::aes(x=cyl, y=mpg)) +
+    ggplot2::geom_point() +
+    ggplot2::labs(title = "title2", subtitle = "subtitle2", tag = "tag2", caption = "footnote2")
+
+  temp_dir_path <- file.path(rprojroot::is_testthat$find_file(), "tempdir")
+  if (!dir.exists(temp_dir_path)){
+    dir.create(file.path(rprojroot::is_testthat$find_file(), "tempdir"))
+  }
+
+  # list of pngs
+  docorator <- as_docorator(
+    x = list(png_obj1, png_obj2),
+    header = fancyhead(fancyrow(center = "first line header"), fancyrow(center = "second line header")),
+    footer = NULL,
+    display_name = "my_first_list",
+    display_loc = temp_dir_path,
+    save_object = FALSE
+  )
+
+  # list of ggplots
+  docorator2 <- as_docorator(
+    x = list(ggplot1, ggplot2),
+    header = fancyhead(fancyrow(center = "first line header"), fancyrow(center = "second line header")),
+    footer = NULL,
+    display_name = "my_first_ggplot_list",
+    display_loc = temp_dir_path,
+    save_object = FALSE
+  )
+
+  # warnings as gt cannot handle rtf figures yet
+  res <- suppressWarnings(suppressMessages( docorator |> render_rtf()
+  ))
+  res2 <- suppressWarnings(suppressMessages( docorator2 |> render_rtf()
+  ))
+
+  expect_true(file.exists(file.path(temp_dir_path, "my_first_list.rtf")))
+  expect_true(file.exists(file.path(temp_dir_path, "my_first_ggplot_list.rtf")))
 
   unlink(temp_dir_path, recursive = TRUE)
 })