+++ NEWS.md
@@ -1,6 +1,8 @@
 # docorator development version
 Improvements
 * Add `qmd` PDF engine
+* Add RTF functionality for PNG and ggplot2 ready for when functionality becomes available in gt with `prep_obj_rtf`
+* Rename `prep_obj` to `prep_obj_tex` and change input to docorator object
 
 Bug fixes
 * Fix issue where `fancyrow()` did not reject string vectors of length > 1 as arguments